fix: the incoming section will now show the function that directly calls the clicked function
It is found that even the official VSCode hierarchy tool also has the bug (When clicking on the `factorial` on line 7 in core.py (it is called by the function `multiply`), the analyser is based on the `factorial` function definition on line 4 and hence the incoming call in callHierarchy.json shows the `main` function in main.py but not the `multiply` function in core.py). The direct caller of the `factorial` function has to be found manually.